Fix spurious global-strict lint problems
I'm not sure why I need this, as I don't see it on every project (for
example, I am not seeing this in the Hextrapolate repo), but this issue
suggests both "global-strict" and "strict" might be needed:

  https://github.com/eslint/eslint/issues/2296

Not sure why, but in any case, it works. Without this hack, we see
problems like this:

> error  Rule 'global-strict' was removed and replaced by: strict  global-strict

(This is confusing, because this is a deprecation notice for
"global-strict", which doesn't appear anywhere in the project.)

With it, the error goes away. If you forget to include `"use strict"`
directive, you correctly get warned:

> error  Use the global form of "use strict"  strict

So, puzzling, but whatever. This gives us a clean lint.